Resolve more PR remarks
